lima: disable mipmapping for linear resources
Due to hardware limitations, it's not possible to implement mipmapping
for linear textures without smart tricks or significantly increasing
memory usage.

RT dimensions have to be a multiply of 16 (tile size) in order for the
textures to be renderable, thus for linear textures we have to specify
stride. The issue is that texture descriptor allows to specify stride
only for all levels at once.

Until we have a better solution, just disable mipmapping for
linear textures. It prevents rendering artefacts and PP MMU faults due
to out of bounds reads for higher levels.

Signed-off-by: Vasily Khoruzhick <anarsoul@gmail.com>