pan/clc: allow building on hosts without libdrm
When using a non-Linux host to cross-compile the panfrost driver we
need to make sure that the 'panfrost_compile' tool can be built without
requiring libdrm or using functions provided by xf86drm.h, which is the
case for panfrost/lib/kmod.

Remove those parts from libpanfrost if only building 'panfrost_compile'
as they are anyway only required for run-time feature detection and
are not used by 'panfrost_compile'.

Signed-off-by: Daniel Golle <daniel@makrotopia.org>