vulkan/meta: Delay sampler creation until do_blit()
For heaps we'll want to use an embedded sampler so we don't want to
cfreate a VkSampler in that case.  This will be easier to handle in
do_blit().  Also, add it to the blit key because we'll need it in the
key in order to use embedded samplers for heaps.